+++ test/unit/index.ts
@@ -7,5 +7,7 @@ import './manage/install.test.ts'
 import './manage/build.test.ts'
 import './manage/build_and_install.test.ts'
 import './manage/diagnostics.test.ts'
+import './manage/detection.test.ts'
+import './manage/mcp_disable_autodetect.test.ts'
 
 console.log('Unit tests loaded.')